Support for 64-bit box lengths looked broken.

+++ include/exiv2/bmffimage.hpp
@@ -92,8 +92,8 @@ namespace Exiv2
           @param start offset in file (default, io_->tell())
          @
          */
-        void parseTiff(uint32_t root_tag, uint32_t length);
-        void parseTiff(uint32_t root_tag, uint32_t length,uint32_t start);
+        void parseTiff(uint32_t root_tag, uint64_t length);
+        void parseTiff(uint32_t root_tag, uint64_t length,uint64_t start);
         //@}
 
         //@{
@@ -103,7 +103,7 @@ namespace Exiv2
           @param start offset in file
          @
          */
-        void parseXmp(uint32_t length,uint32_t start);
+        void parseXmp(uint64_t length,uint64_t start);
         //@}
 
         //! @name Manipulators
